Added dialog to user events on all list screens
Previously the edit/delete dialog only showed when calling manage
events. Now works from any list as long as the currently signed in
user's event is clicked.
Note: doing this to prepare to get rid of admin panel (most of those
things aren't needed or barely used and can be moved to settings)